refactor: package 변경
- config -> core/config로 변경
